+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/svelte
 
+## 3.0.42
+
+### Patch Changes
+
+- Updated dependencies [de5c066]
+  - ai@5.0.42
+
+## 3.0.41
+
+### Patch Changes
+
+- Updated dependencies [cd91e4b]
+  - ai@5.0.41
+
 ## 3.0.40
 
 ### Patch Changes